requirements.txt
# Core AI and Vector Database
openai>=1.3.0
chromadb>=0.4.15
tiktoken>=0.5.1

# Web Framework and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Requests and Validation
pydantic>=2.0.0
requests>=2.31.0

# Optional: Enhanced development experience
rich>=13.0.0
colorama>=0.4.6

# Optional: Additional utilities
python-multipart>=0.0.6
aiofiles>=23.0.0